requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
django_filter
python-dotenv
coverage